+++ host/usb/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Fixed
 
+- Fixed race between root port disconnect handling and device recycle path that could abort with `ESP_ERR_NOT_FOUND` from `dev_tree_node_dev_gone()` (https://github.com/espressif/esp-idf/issues/18366)
 - List also suspended devices in `usb_host_device_addr_list_fill()` function
 
 ## [1.4.0] - 2026-04-02